add policies and implementation

+++ app/Http/Livewire/FakturService/Index.php
@@ -28,6 +28,9 @@ class Index extends Component
 
     public function destroy(FakturService $fakturService)
     {
+        $this->authorize('delete', $fakturService);
+
+
         $fakturService->delete();
     }
