+++ app/Http/Livewire/JenisService/Index.php
@@ -3,12 +3,14 @@
 namespace App\Http\Livewire\JenisService;
 
 use App\Models\JenisService;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class Index extends Component
 {
     use WithPagination;
+    use AuthorizesRequests;
 
     // Search and sort
     public $query = "";
@@ -28,6 +30,8 @@ class Index extends Component
 
     public function destroy(JenisService $jenisService)
     {
+        $this->authorize('delete', $jenisService);
+
         $jenisService->delete();
     }
